+++ LT.DigitalOffice.Kernel.BrokerSupport.nuspec
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <package >
   <metadata>
-    <id>LT.DigitalOffice.Kernel.RedisSupport</id>
+    <id>LT.DigitalOffice.Kernel.BrokerSupport</id>
     <version>1.0.0</version>
-    <title>LT.DigitalOffice.Kernel.RedisSupport</title>
+    <title>LT.DigitalOffice.Kernel.BrokerSupport</title>
     <authors>LT Digital Office</authors>
     <owners>LT Digital Office</owners>
     <projectUrl>https://github.com/LT-Students/DigitalOffice-Kernel</projectUrl>
@@ -18,13 +18,16 @@
     <tags>DigitalOffice</tags>
     <dependencies>
       <group targetFramework="net5.0">
-        <dependency id="StackExchange.Redis" version="2.2.50" />
+        <dependency id="MassTransit" version="7.1.7"/>
+        <dependency id="MassTransit.AspNetCore" version="7.1.7"/>
+        <dependency id="MassTransit.Extensions.DependencyInjection" version="7.1.7"/>
+        <dependency id="MassTransit.RabbitMQ" version="7.1.7"/>
         <dependency id="LT.DigitalOffice.Kernel" version="1.1.51" />
       </group>
     </dependencies>
   </metadata>
   <files>
-    <file src="artifacts\kernel_redis\LT.DigitalOffice.Kernel.RedisSupport.dll" target="lib\net5.0" />
-    <file src="artifacts\kernel_redis\LT.DigitalOffice.Kernel.RedisSupport.pdb" target="lib\net5.0" />
+    <file src="artifacts\kernel_broker\LT.DigitalOffice.Kernel.BrokerSupport.dll" target="lib\net5.0" />
+    <file src="artifacts\kernel_broker\LT.DigitalOffice.Kernel.BrokerSupport.pdb" target="lib\net5.0" />
   </files>
 </package>
