+++ src/Kernel.BrokerSupport/Attributes/ParseEntity/IgnoreParseAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace LT.DigitalOffice.Kernel.BrokerSupport.Attributes.ParseEntity
+{
+  [AttributeUsage(AttributeTargets.Property)]
+  public class IgnoreParseAttribute : Attribute
+  {
+  }
+}
