+++ src/Kernel.BrokerSupport/Attributes/ParseEntity/Models/Requests/IFindParseEntitiesRequest.cs
@@ -1,0 +1,10 @@
+﻿namespace LT.DigitalOffice.Kernel.BrokerSupport.Attributes.ParseEntity.Models.Requests
+{
+  public interface IFindParseEntitiesRequest
+  {
+    static object CreateObj()
+    {
+      return new();
+    }
+  }
+}
