+++ src/Kernel.BrokerSupport/Attributes/ParseEntity/ParseEntityAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace LT.DigitalOffice.Kernel.BrokerSupport.Attributes.ParseEntity
+{
+  [AttributeUsage(AttributeTargets.Class)]
+  public class ParseEntityAttribute : Attribute
+  {
+  }
+}
