+++ src/Kernel.ImageSupport/Kernel.ImageSupport.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <Version>1.0.0</Version>
+
+    <TargetFramework>net5.0</TargetFramework>
+    <AssemblyName>LT.DigitalOffice.Kernel.ImageSupport</AssemblyName>
+    <RootNamespace>LT.DigitalOffice.Kernel.ImageSupport</RootNamespace>
+    <IsPackable>true</IsPackable>
+    <IncludeSymbols>true</IncludeSymbols>
+    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
+    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
+
+    <GenerateDocumentationFile>false</GenerateDocumentationFile>
+
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="LT.DigitalOffice.Kernel" Version="1.1.47" />
+    <PackageReference Include="SixLabors.ImageSharp" Version="1.0.4" />
+    <PackageReference Include="Svg" Version="3.3.0" />
+  </ItemGroup>
+
+</Project>
