+++ src/Kernel.RedisSupport/Kernel.RedisSupport.csproj
@@ -18,7 +18,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="LT.DigitalOffice.Kernel" Version="1.1.50" />
+      <PackageReference Include="LT.DigitalOffice.Kernel" Version="1.1.51" />
       <PackageReference Include="StackExchange.Redis" Version="2.2.50" />
     </ItemGroup>
