--- src/Kernel/Attributes/ParseEntity/IgnoreParseAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace LT.DigitalOffice.Kernel.Attributes.ParseEntity
-{
-    [AttributeUsage(AttributeTargets.Property)]
-    public class IgnoreParseAttribute : Attribute
-    {
-    }
-}
