--- src/Kernel/Attributes/ParseEntity/Models/Requests/IFindParseEntitiesRequest.cs
@@ -1,10 +1,0 @@
-﻿namespace LT.DigitalOffice.Kernel.Middlewares.ParseEntities.Models.Requests
-{
-    public interface IFindParseEntitiesRequest
-    {
-        static object CreateObj()
-        {
-            return new();
-        }
-    }
-}
