--- src/Kernel/Attributes/ParseEntity/ParseEntityAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace LT.DigitalOffice.Kernel.Attributes.ParseEntity
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public class ParseEntityAttribute : Attribute
-    {
-    }
-}
